Moved Flask app to library and made optional.
The Flask app seems to be causing some trouble anyway, so I decided
to make it optional. Everything ostensibly works, it just has
some ugly output whenever you quit it while its running.

The signal handlers were also getting in the way of each other
between the Flask server and the one I installed. Now they're
mutually exclusive.
