+++ main.py
@@ -1,80 +1,22 @@
 """main.py
 Simple flask application for streaming motion jpg images to a web browser.
 """
-from flask import Flask, render_template, Response, request, render_template
 from camera.processor import VideoProcessor
 from camera.camera import VideoCamera
 import cv2
-import requests
 import signal
-from threading import Thread
 from gui.pitft_main_window import PiTftMainWindow
-
-app = Flask(__name__)
-"""The global Flask app used for talking to clients through HTTP"""
-
-
-@app.route('/')
-def index():
-    """Setup the index page when people navigate to the machine name and port"""
-    return render_template('index.html')
-
-
-def gen(vid_proc):
-    """ Generate a new image stream from the camera object.
-
-    Args:
-        vid_proc(VideoProcessor): The VideoProcessor object to grab the camera
-        frame and perform image processing on it.
-
-    Yields:
-        A new image frame in motion jpg format
-    """
-    count = 0
-    while True:
-        frame = vid_proc.get_last_frame()
-        if frame is not None:
-            yield(b'--frame\r\n'
-                  b'Content-Type:image/jpeg\r\n\r\n' +
-                  cv2.imencode('.jpg', frame)[1].tostring() +
-                  b'\r\n\r\n')
-
-
-@app.route('/video_feed')
-def video_feed():
-    """ Start a new video feed by telling the browser we're about to send motion
-    jpg frames
-
-    Returns:
-        The Flask response to the client.
-    """
-    return Response(gen(vid_proc),
-                    mimetype='multipart/x-mixed-replace; boundary=frame')
-
-
-def start_flask_thread():
-    """Wrapper for starting the Flask app in a different thread."""
-    print("Starting the Flask app.")
-    # app.run(host='0.0.0.0', debug=False, use_reloader=False)
-    app.run(host='0.0.0.0')
-
-
-@app.route('/shutdown', methods=['POST'])
-def shutdown():
-    """Wrapper to handles shutting down the server with a request"""
-    shutdown_flask_server()
-    return "Flask server shutting down..."
-
-
-def shutdown_flask_server():
-    """Handle stopping the server by running the server's shutdown script."""
-    func = request.environ.get('werkzeug.server.shutdown')
-    if func is None:
-        raise RuntimeError("Not running with Werkzeug Server")
-    func()
+from flask_app.flask_app import FlaskApp
+from argparse import ArgumentParser
 
 
 if __name__ == '__main__':
+    parser = ArgumentParser(description="PatrolBot App.")
+    parser.add_argument('--flask', action='store_const', const=True,
+                        help="Use Flask to enable viewing imagery from the "
+                             "browser on port 5000.")
+    args = parser.parse_args()
+
     # The main PyGame window that we will use to display and get touch
     # screen input from
     mw = PiTftMainWindow((320, 240))
@@ -88,20 +30,18 @@ if __name__ == '__main__':
 
     # Get one frame so that the video fee will populate
     vid_proc.get_frame()
-    # Kick off the viewer in another thread to get it out of the way
-    flask_thread = Thread(target=start_flask_thread, args=())
-    flask_thread.start()
+    if args.flask:
+        flask_app = FlaskApp(vid_proc)
+    else:
+        # Install our own signal handler so the user can gracefully exit.
+        def signal_handler(sig, frame):
+            global running
+            running = False
+            print("\nPatrolBot: Caught SIGINT. Quitting.")
 
-    def signal_handler(sig, frame):
-        global running
-        running = False
-        print("\nPatrolBot: Caught SIGINT. Quitting.")
-        requests.post("http://localhost:5000/shutdown")
-        flask_thread.join(timeout=1)
+        signal.signal(signal.SIGINT, signal_handler)
 
-    signal.signal(signal.SIGINT, signal_handler)
     running = True
-
     while running:
         vid_proc.get_frame(show_fps=True)
         mw.update(
